chore: migrate home-page to rsc (#478)
* move global

* add next/third-parties

* layout

* move 404 and error

* move index

* upgrade next to latest

* global error with sentry

* remove app

* typed routes

* same react types version

* remove font import

* home page renders
