+++ package.json
@@ -30,6 +30,7 @@
     "@aws-sdk/s3-request-presigner": "^3.427.0",
     "@headlessui/react": "^2.0.4",
     "@heroicons/react": "^2.1.3",
+    "@next/third-parties": "^14.2.4",
     "@paralleldrive/cuid2": "^2.2.2",
     "@prisma/client": "^5.12.1",
     "@sentry/nextjs": "^8.5.0",
@@ -40,12 +41,12 @@
     "handlebars": "^4.7.8",
     "jszip": "^3.10.1",
     "marked": "^12.0.2",
-    "next": "^14.2.2",
+    "next": "^14.2.4",
     "next-auth": "5.0.0-beta.19",
     "nodemailer": "^6.9.13",
     "qs": "^6.12.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "sanitize-html": "^2.13.0",
     "sharp": "0.32.6",
     "swiper": "11.0.3",
@@ -66,7 +67,7 @@
     "@types/node": "^20.12.12",
     "@types/nodemailer": "^6.4.15",
     "@types/qs": "^6.9.15",
-    "@types/react": "18.2.0",
+    "@types/react": "18.3.1",
     "@types/sanitize-html": "^2.11.0",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
