+++ src/components/Filter/Search.tsx
@@ -1,13 +1,16 @@
+'use client'
+
 import { MagnifyingGlassIcon } from '@heroicons/react/20/solid'
-import { useRouter } from 'next/router'
+import { usePathname, useRouter } from 'next/navigation'
 
 import { Input, InputGroup } from '@/components/ui'
 import { useQueryParams } from '@/lib/hooks/useQueryParams'
 
 export const Search = () => {
   const router = useRouter()
+  const pathname = usePathname()
   const [query, setQuery] = useQueryParams()
-  const defaultValue = new URLSearchParams(query).get('search')
+  const defaultValue = new URLSearchParams(query ?? '').get('search')
 
   return (
     <form
@@ -15,20 +18,6 @@ export const Search = () => {
       onSubmit={(e) => {
         e.preventDefault()
         const formData = new FormData(e.currentTarget)
-
-        if (!formData.get('search')) {
-          router.push({
-            ...router.query,
-            search: null,
-          })
-          return
-        }
-        router.push(
-          setQuery({
-            ...router.query,
-            search: formData.get('search'),
-          })
-        )
       }}
     >
       <InputGroup>
