+++ src/components/ui/Dropdown.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import * as Headless from '@headlessui/react'
 import type React from 'react'
