+++ src/components/ui/Table.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type React from 'react'
 import { createContext, useContext, useState } from 'react'
